@@ -1,10 +1,10 @@
 module DDatHome-go
 
-go 1.14
+go 1.16
 
 require (
-	github.com/kardianos/service v1.0.0
-	github.com/stretchr/testify v1.5.1 // indirect
-	github.com/tidwall/gjson v1.6.0
-	golang.org/x/net v0.0.0-20200520182314-0ba52f642ac2
+	github.com/goccy/go-json v0.9.10
+	github.com/kardianos/service v1.2.1
+	golang.org/x/net v0.0.0-20220802222814-0bcc04d9c69b
+	golang.org/x/sys v0.0.0-20220731174439-a90be440212d // indirect
 )
